requirements.txt
argparse==1.2.1
clint==0.3.1
paul==0.0.1
psutil==1.0.1
python-gnupg==0.3.4
texttable==0.8.1
wsgiref==0.1.2
xmlrpclib==1.0.1
